refactor(dsh-mcp-manager): 阶段4工作空间路由域成形——src/workspace/ + interface.ts + 调用点全切（Part of #664）
Summary:
- 新建 src/workspace/（D10 门面 interface.ts）：root-resolution.ts（findProjectRoot/
  normalizedProjectRoot 自 manager.ts 方法化迁出 + makeResolveRoot 自 apply-runtime.ts
  迁入，B3 修复落点）、full-name.ts（fullServerName/parseFullServerName/
  bareServerName/normalizeToolName 自 middleware-utils.ts）、scope.ts
  （SCOPE_GLOBAL/SCOPE_PROJECT/normalizeScope）、middleware-mode.ts
  （normalizeMiddlewareMode 自 middleware-const.ts 只迁函数）、constants.ts
  （MIDDLEWARE_GLOBAL_ROOT 双源收敛单源）
- 旧位置删除定义保留 lib 导出面：middleware.ts / index.ts re-export 链改从
  workspace/interface.ts 取；scope.ts 文件删除（引用全切）；middleware-const.ts
  仅留 limits/超时常量（阶段 6 归 connection/runtime/limits.ts）
- 静态面同步：mutation-topology supervisor 段移除 src/scope.ts（C4 变异盲区
  明示），gen-stryker-conf 重生成；unit-manager2 findProjectRoot 测试改纯函数直测

Design decisions:
- 迁移模式=新文件承载实现+旧位置删定义，不 git mv（阶段 6 集中搬移）
- 薄转发不做双份：scope.ts 全切后删除；makeResolveRoot 迁出 apply-runtime.ts
- lib 导出面零删除，新增 findProjectRoot/normalizedProjectRoot/makeResolveRoot
  三个 workspace 域符号（T1 测试面）

Related issue #664
