+++ packages/dsh-mcp-manager/src/supervisor.ts
@@ -11,7 +11,7 @@
 import { createHash } from "node:crypto";
 import { createTransport } from "./transport.ts";
 import type { StdioTransport, HttpTransport } from "./transport.ts";
-import { SCOPE_GLOBAL, SCOPE_PROJECT } from "./scope.ts";
+import { SCOPE_GLOBAL, SCOPE_PROJECT } from "./workspace/interface.ts";
 import { MCPClient } from "./protocol.ts";
 import { defaultCallResultFallbackText, projectCallToolResult, createRedactor, msgOf } from "./pipeline/interface.ts";
 import { RECONNECT_DEFAULTS, resolveReconnect, type ReconnectPolicy } from "./connection/interface.ts";
